feat: enables the detailed control of terminals which may differ for different n-gram components (except the first one).
Control happens by passing one or several TerminalSymbol enum members (or equivalent strings), or one or several arbitrary terminal symbols. This commit also introduces the setting default_terminal_symbol, which is the one used for TerminalSymbol.DEFAULT and can be changed by the user. Caveat: Any row where all columns of a component have missing values is considered a terminal, even if it is the result from missing values in a preceding n-gram's first component. While this facilitated the implementation quite a bit, more thought should be put in the NgramAnalyzers on how to deal with mid-sequence "holes" in the first place, because the current behaviour is probably inappropriate (where the missing values are treated as values when they are the first n-gram component, but as a terminal otherwise). Simply dropping them, however, may lead to wrong, inexisting transitions. Probably more though needs to go into this on a case-to-case basis for individual features.